pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "atpoe"
version = "0.1.0"
description = "Admitting the Possibilities of Error - Curve Generation"
authors = [
    {name = "AtPoE Project"},
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "streamlit",
    "Pillow",
    "matplotlib",
    "PyYAML",
    "pytest",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "flake8",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["atpoe*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

